Pipeline refactor 8: full pipeline tracking
